+++ forui/lib/src/widgets/accordion/accordion_item.dart
@@ -82,8 +82,8 @@ class _FAccordionItemState extends State<FAccordionItem> with TickerProviderStat
   void didChangeDependencies() {
     super.didChangeDependencies();
 
-    final FAccordionItemData(:index, :controller, :style) = FAccordionItemData.of(context);
-    controller.removeItem(index);
+    final InheritedAccordionData(:index, :controller, :style) = InheritedAccordionData.of(context);
+    controller.remove(index);
 
     _controller?.dispose();
     _body?.dispose();
@@ -97,14 +97,21 @@ class _FAccordionItemState extends State<FAccordionItem> with TickerProviderStat
     _body = CurvedAnimation(curve: style.expandCurve, reverseCurve: style.collapseCurve, parent: _controller!);
     _icon = Tween<double>(begin: 0, end: 0.5).animate(_body!);
 
-    if (!controller.addItem(index, _controller!)) {
+    if (!controller.add(index, _controller!)) {
       throw StateError('Number of expanded items must be within the min and max.');
     }
   }
 
   @override
+  void dispose() {
+    _body?.dispose();
+    _controller?.dispose();
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
-    final FAccordionItemData(:index, :controller, style: inheritedStyle) = FAccordionItemData.of(context);
+    final InheritedAccordionData(:index, :controller, style: inheritedStyle) = InheritedAccordionData.of(context);
     final style = widget.style ?? inheritedStyle;
 
     return Column(
@@ -155,12 +162,5 @@ class _FAccordionItemState extends State<FAccordionItem> with TickerProviderStat
         FDivider(style: style.dividerStyle),
       ],
     );
-  }
-
-  @override
-  void dispose() {
-    _body?.dispose();
-    _controller?.dispose();
-    super.dispose();
   }
 }