+++ samples/lib/widgets/accordion.dart
@@ -6,9 +6,8 @@ import 'package:forui/forui.dart';
 import 'package:forui_samples/sample.dart';
 
 final controllers = {
-  'default': FAccordionController(),
-  'default-max': FAccordionController(max: 2),
-  'radio': FAccordionController.radio(),
+  'default': FAccordionController(max: 1),
+  'max': FAccordionController(max: 2),
 };
 
 @RoutePage()