+++ forui/CHANGELOG.md
@@ -1,5 +1,16 @@
 ## 0.15.0
 
+### `FAccordionController`
+
+* **Breaking** Make `FAccordionController.controllers` private.
+* **Breaking** Remove `FAccordionController.radio(...)` - use `FAccrdionController(max: 1)` instead.
+* **Breaking** Remove `FAccordionController.validate(...)`.
+
+* Fix `FAccordionController.expand(...)` wrongly returning false if the item is already expanded.
+* Fix `FAccordionController.collapse(...)` wrongly returning false if the item is already collapsed.
+
+
+
 ### `FMultiSelect` (new)
 * Add `FMultiSelect`.
 * Add `FMultiSelectController`.